feat: Update

+++ utils.py
@@ -9,8 +9,8 @@ def loadConfig(current_version):
     with open("config.json") as config_file:
         config = json.load(config_file)
         try:
-            if int(config['version']) < int(current_version):
-                if int(config['version']) == 1.0:
+            if float(config['version']) < float(current_version):
+                if float(config['version']) == 1.0:
                     config.update({"fullscreen":"Windowed"})
                 config['version'] = current_version
                 saveConfig(config)
